backend/requirements.txt
django
djangorestframework
sentence-transformers
faiss-cpu
openai
numpy
google-genai
pypdf
python-docx
djangorestframework-simplejwt
requests
dotenv
beautifulsoup4
dj-database-url
psycopg[binary]
trafilatura
gunicorn
whitenoise
django-cors-headers
deepeval
celery[redis]
redis
youtube-transcript-api
google-auth
cryptography